Don't add controller node.
What happens if we don't add the controller node to the interface?

For this node server, the controller node is nothing but a parent
for the real device nodes.  It doesn't do any actual control of
anything. In theory, we can still use the node class, but just not
add it to the interface and don't create each player node as a child.

If this works, then there's no need for the controller to be a Node
class, it can just be an internally defined class based on Object.

My one concern is with the START event.  Since that looks for a
node address to send the event to, will it continue to work?  We may
have to call start() directly from init.